moved close_on_esc and added window_id to touch events
