+++ crates/bevy_core/Cargo.toml
@@ -17,6 +17,8 @@ bevy_math = { path = "../bevy_math", version = "0.9.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.9.0-dev", features = ["bevy"] }
 bevy_tasks = { path = "../bevy_tasks", version = "0.9.0-dev" }
 bevy_utils = { path = "../bevy_utils", version = "0.9.0-dev" }
+bevy_window = { path = "../bevy_window", version = "0.9.0-dev" }
+bevy_input = { path = "../bevy_input", version = "0.9.0-dev" }
 
 # other
 bytemuck = "1.5"
