+++ crates/bevy_input/Cargo.toml
@@ -19,6 +19,7 @@ bevy_ecs = { path = "../bevy_ecs", version = "0.9.0-dev" }
 bevy_math = { path = "../bevy_math", version = "0.9.0-dev" }
 bevy_utils = { path = "../bevy_utils", version = "0.9.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.9.0-dev" }
+bevy_window = { path = "../bevy_window", version = "0.9.0-dev" }
 
 # other
 serde = { version = "1", features = ["derive"], optional = true }
