+++ crates/bevy_window/Cargo.toml
@@ -19,8 +19,6 @@ bevy_ecs = { path = "../bevy_ecs", version = "0.9.0-dev" }
 bevy_math = { path = "../bevy_math", version = "0.9.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.9.0-dev" }
 bevy_utils = { path = "../bevy_utils", version = "0.9.0-dev" }
-# Used for close_on_esc
-bevy_input = { path = "../bevy_input", version = "0.9.0-dev" }
 raw-window-handle = "0.4.2"
 
 # other
