+++ crates/bevy_window/src/system.rs
@@ -1,8 +1,7 @@
-use crate::{Window, WindowCloseRequested, WindowFocused, WindowId, Windows};
+use crate::{Window, WindowCloseRequested, Windows};
 
 use bevy_app::AppExit;
 use bevy_ecs::prelude::*;
-use bevy_input::{keyboard::KeyCode, Input};
 
 /// Exit the application when there are no open windows.
 ///
@@ -30,28 +29,5 @@ pub fn close_when_requested(
 ) {
     for event in closed.iter() {
         windows.get_mut(event.id).map(Window::close);
-    }
-}
-
-/// Close the focused window whenever the escape key (<kbd>Esc</kbd>) is pressed
-///
-/// This is useful for examples or prototyping.
-pub fn close_on_esc(
-    mut focused: Local<Option<WindowId>>,
-    mut focused_events: EventReader<WindowFocused>,
-    mut windows: ResMut<Windows>,
-    input: Res<Input<KeyCode>>,
-) {
-    // TODO: Track this in e.g. a resource to ensure consistent behaviour across similar systems
-    for event in focused_events.iter() {
-        *focused = event.focused.then_some(event.id);
-    }
-
-    if let Some(focused) = &*focused {
-        if input.just_pressed(KeyCode::Escape) {
-            if let Some(window) = windows.get_mut(*focused) {
-                window.close();
-            }
-        }
     }
 }
