+++ examples/2d/rotation.rs
@@ -16,7 +16,7 @@ fn main() {
                 .with_system(snap_to_player_system)
                 .with_system(rotate_to_player_system),
         )
-        .add_system(bevy::window::close_on_esc)
+        .add_system(bevy::core::close_on_esc)
         .run();
 }
