+++ examples/games/alien_cake_addict.rs
@@ -34,7 +34,7 @@ fn main() {
                 .with_run_criteria(FixedTimestep::step(5.0))
                 .with_system(spawn_bonus),
         )
-        .add_system(bevy::window::close_on_esc)
+        .add_system(bevy::core::close_on_esc)
         .run();
 }
