+++ examples/window/multiple_windows.rs
@@ -10,7 +10,7 @@ fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
         .add_startup_system(setup)
-        .add_system(bevy::window::close_on_esc)
+        .add_system(bevy::core::close_on_esc)
         .run();
 }
